Move the JSON Schema inference types into src/types/json.d.ts
index.d.ts carried the FromJSONSchema engine inline; it now sits beside
the dialect and Standard Schema mirrors as src/types/json.d.ts. The JSON
data type moves with it so the new file needs no import back into
index.d.ts, which imports and re-exports it the same way it already does
for the other two.

Behavior-neutral: every inferred type string is unchanged and the five
affected specs each drop one instantiation.
