+++ packages/sury/specs/jsonschema-empty-range.yaml
@@ -3,7 +3,7 @@ ts:
   schema: 'S.fromJSONSchema({ type: "number", minimum: 5, maximum: 1 })'
   input: number
   output: number
-  instantiations: 388
+  instantiations: 387
 vs:
   zod:
     _skip: not-applicable
