+++ packages/sury/specs/jsonschema-int-maximum-above-int32.yaml
@@ -13,7 +13,7 @@ ts:
   schema: 'S.fromJSONSchema({ type: "integer", maximum: 3000000000 })'
   input: number
   output: number
-  instantiations: 388
+  instantiations: 387
 vs:
   zod:
     _skip: not-applicable
