+++ packages/sury/specs/jsonschema-int-minimum-above-int32.yaml
@@ -12,7 +12,7 @@ ts:
   schema: 'S.fromJSONSchema({ type: "integer", minimum: 3000000000 })'
   input: number
   output: number
-  instantiations: 388
+  instantiations: 387
 vs:
   zod:
     _skip: not-applicable
